Use HashSet to track Uses in a Def.
If we continue using a `Vec` with the Use also
storing an index into this vec, removal becomes tricky.
If we need to remove a Use A, and move the last Use B in
the `uses`vec to it's place, then we must update B to
use the index in its `Use` in the operand. Now, if A and B
are both in the same Operation, the update would be tricky
without being able to get two (interior) mutable references.