IR format printers (#22)
Backport from #20 

This PR introduces the `irfmt/printers` module with common printer functions.

In #20 we implement `asm_format` directives as macros. These macros will  use the printer functions defined here (and more to come). As secondary goal we want to make it easier for downstream developers to implement
their own printers, while trying to keep some kind of standardized format. For this reason we also implement Printable for some more IR objects in pliron.

The change also removes the `PrintableIter` trait in favor of the `list_with_sep` and `iter_with_sep` printer functions in the irfmt package. This change ensures that we have a single trait for printing in the library.

As discussed in #21 we do not require developers to add the type_id or
attribute_id manually to the printer anymore. `pliron` now implements a
printer for TypeObj/AttribObj that will always print the IDs.

---------

Co-authored-by: urso <steffen.siering at gmail.com>
Co-authored-by: Vaivaswatha N <vaivaswatha@users.noreply.github.com>