Move parser helpers to irfmt::parsers (#29)
This is an intermediary step backporting some changes from #20 into
main.

This PR focuses on moving the existing parsers into the `irfmt::parsers`
module. Idea is to have parsers (and helpers) under the `irfmt` module
for discoverability, similar to the `irfmt::printers` module.

In addition to that I implemented the `Parsable` trait for `AttrObj` and
`Ptr<TypeObj>` by copying the former `attr_parse` and `type_parse`
function as is into the trait. I did this change to improve symmetry: If
a type implements `Printable`, then it should implement `Parsable` as
well.

Besides implementing `Parsable` for the 2 types there is no change in
the implementation logic. Just moving symbols around.

Co-authored-by: urso <steffen.siering at gmail.com>