Add a simple preconfigured walker (#32)
This PR provides a set of walkers (one for each IR node kind - op, block
and region) that are pre-configured (when called with the root node) on
when the callback is called (w.r.t walking over the children) and the
order in which children are walked over.

The callback itself is a simple common type for all IR node kinds.

The walker provides `&mut Context` to the callback. What's safe and not
is documented in the module docs.

Another slightly more extensible alternate to this approach is to
provide a trait that combines the configuration, state and callbacks
into one. (i.e., the trait has query functions to get each). This will
allow the flexibility of objects to change the configuration and
callbacks *during* the walk. I don't know if that's useful, but it's
doable. It's likely going to make usage more verbose, so I suppose we're
better off with the current version.

References: @urso's
[proposals](https://github.com/vaivaswatha/pliron/discussions/25#discussioncomment-8364083)
and my previous [PR](#31).