Ensure super-interface verifiers run before sub-interface verifiers.
This change introduces a `TypeId` (of the `dyn Interface`) based
dependence tracker for interfaces (both `Op` and `Attribute`) to
express super-interfaces. While Rust's trait system already tracks
this, that isn't good enough for us to ensure that a super-interface's
verifier is run before its sub-interface's verifier.

We now have `decl_(op/attr)_interface!` macros to specify new
interfaces for `Op`s and `Attribute`s.

As a happy benefit, we don't require the `inventory` crate anymore,
which we were using to track all the interface verifiers for an
`Op` or `Attribute`.