Ignore leading/trailing whitespaces during parsing of literals
This allows specifying spaces as part of literals in `#[format]`
strings, allowing for a prettier formatting.

This was needed because the parser we generate anyway parses
and ignores whitespaces before and after a literal, thus leading
to an error if an actual whitespace needs to be parsed. So if an
actual whitespace needs to be parsed, we trim that.