Remove `parent_for_regions` from the parser state
This was incorrect. When we enter a region (during parsing), the
decision to create a new scope or not depends on whether the parent
op is "IsolatedFromAbove". Having a common parent Op that does not
implement "IsolatedFromAbove" was incorrect. We solve the problem
of "requiring a parent to be parsed before parsing its regions"
in two ways now:

1. The canonical syntax parser anyway has regions at the end of
an operation, so the entire operation is parsed and created before
parsing the region and passing the operation pointer to it during.
2. The parsers derived from `#[format(..)]` create a temporary
Operation with the same `OpId` as the one being parsed, passes
that to the region parsers and then transfers the regions to
their final parent after that is created. We also take care
to create this temporary parent only when the syntax specifies
that the Op has a region, to avoid unnecessarily creating and
erasing these temporary Ops.