Introduce logging via the `log` and `env_logger` crates
The llvm-opt tool is now a separate crate, so that only
binary (and test) crates need the `env_logger` dependence,
and also set the `log` crate to `release_max_level_warn`.

Crates that use `pliron` that want to statically disable logging
will now need to also add `log` and set the feature flag to
disable logging. This is an unfortunate side-effect of the feature
flag being global.

See the `log` and `env_logger` crates' docs for usage details.