Make `Operation::get_op works` fast.
Wrapping a `Ptr<Operation>` inside an `Op` required
looking up the wrapper function by indexing (based on `OpId`)
into a table in `Context`. This was slow. So instead,
store the wrapper function directly within `Operation`.

This changes the signature of `Operation::new`, as it now
takes the wrapping function instead of `OpId` as an argument.